+++ CavrnusSdk/Runtime/Scripts/API/CavrnusTypes/CavrnusAuthentication.cs
@@ -5,10 +5,12 @@ namespace CavrnusSdk.API
 	public class CavrnusAuthentication
 	{
 		internal RestApiEndpoint Endpoint;
+		public string Token;
 
-		internal CavrnusAuthentication(RestApiEndpoint endpoint)
+		internal CavrnusAuthentication(RestApiEndpoint endpoint, string token)
 		{
 			Endpoint = endpoint;
+			Token = token;
 		}
 	}
 }